requirements.txt
## Core ML/AI Dependencies
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
peft>=0.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Multi-modal Model Dependencies
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0

# API and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Memory and Vector Database
chromadb>=0.4.18
redis>=5.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Use faiss-gpu if you have CUDA

# LLM Framework and Tools
langchain>=0.0.340
langchain-community>=0.0.1
openai>=1.3.0
tiktoken>=0.5.0

# UI and Visualization
gradio>=4.8.0
streamlit>=1.29.0  # Alternative to Gradio
plotly>=5.18.0

# Training and Monitoring
wandb>=0.16.0
tensorboard>=2.15.0
datasets>=2.15.0
evaluate>=0.4.0
scikit-learn>=1.3.0

# Utility Libraries
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.0
tqdm>=4.66.0
rich>=13.7.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Optional: For production deployment
gunicorn>=21.2.0
nginx>=1.25.0
docker>=6.1.0
docker-compose>=1.29.0

# Optional: Additional AI Tools
# mcp>=0.1.0  # When MCP Python SDK is available
# autogen>=0.2.0  # For multi-agent scenarios
# guidance>=0.1.0  # For structured generation Core ML/AI Dependencies
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
peft>=0.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Multi-modal Model Dependencies
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0

# API and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Memory and Vector Database
chromadb>=0.4.18
redis>=5.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Use faiss-gpu if you have CUDA

# LLM Framework and Tools
langchain>=0.0.340
langchain-community>=0.0.1
openai>=1.3.0
tiktoken>=0.5.0

# UI and Visualization
gradio>=4.8.0
streamlit>=1.29.0  # Alternative to Gradio
plotly>=5.18.0

# Training and Monitoring
wandb>=0.16.0
tensorboard>=2.15.0
datasets>=2.15.0
evaluate>=0.4.0
scikit-learn>=1.3.0

# Utility Libraries
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.0
tqdm>=4.66.0
rich>=13.7.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Optional: For production deployment
gunicorn>=21.2.0
nginx>=1.25.0
docker>=6.1.0
docker-compose>=1.29.0

# Optional: Additional AI Tools
# mcp>=0.1.0  # When MCP Python SDK is available
# autogen>=0.2.0  # For multi-agent scenarios
# guidance>=0.1.0  # For structured generation